backend/requirements.txt
fastapi==0.115.12
joblib==1.4.2
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
protobuf==6.30.2
pydantic==2.11.3
scikit_learn==1.6.1
seaborn==0.13.2
ta==0.11.0
tensorflow==2.19.0
tensorflow-intel==2.18.0
tweepy==4.15.0
xgboost==3.0.0
yfinance==0.2.54
google-generativeai
google-genai
